Update Docker to be modular and use docker-compose
This commit updates the Dockerfile to separate out redis and mysql, and
adds a docker-compose.yml file for use with docker-compose to build the
containers as a group. Refer to docker-compose documentation for info.

This is not intended for production - to simplify handling of email, all
mail is being sunk to http://127.0.0.1:8001/ via a web UI. This makes it
easier to test where you don't have a valid email server.

Configuration can be adjusted in config/config_docker.json.

Known issues at present:

- Attachments/uploads should be put into a data volume for portability
- The installation thinks the hostname is 0.0, and the redirect after
setup appears to be incorrect.

There are likely other issues - this is more intended as a proof-of-concept.
The (reduced) Dockerfile can likely be further simplified and improved.